+++ integrations/upgrade/js-config.test.ts
@@ -1296,10 +1296,10 @@ describe('border compatibility', () => {
         "
         --- src/index.html ---
         <div
-          class="[width:calc(var(--spacing)*2)]
-            [width:calc(var(--spacing)*4.5)]
+          class="[width:--spacing(2)]
+            [width:--spacing(4.5)]
             [width:var(--spacing-5_5)]
-            [width:calc(var(--spacing)*13)]
+            [width:--spacing(13)]
             [width:var(--spacing-100)]
             [width:var(--spacing-miami)]"
         ></div>
@@ -1332,10 +1332,10 @@ describe('border compatibility', () => {
         }
 
         .container {
-          width: calc(var(--spacing) * 2);
-          width: calc(var(--spacing) * 4.5);
+          width: --spacing(2);
+          width: --spacing(4.5);
           width: var(--spacing-5_5);
-          width: calc(var(--spacing) * 13);
+          width: --spacing(13);
           width: var(--spacing-100);
           width: var(--spacing-miami);
         }
@@ -1515,7 +1515,7 @@ describe('border compatibility', () => {
         @utility container {
           margin-inline: auto;
           padding-inline: 2rem;
-          @media (width >= theme(--breakpoint-sm)) {
+          @media (width >= --theme(--breakpoint-sm)) {
             max-width: none;
           }
           @media (width >= 48rem) {